Added unit-tests for appointments service, refactor documents service(minimal api), change to mongodb ef provider, fix issues

+++ InnoClinic/Services/Profiles/Profiles.Application/Commands/Doctors/DeleteDoctor/DeleteDoctorCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿public class DeleteDoctorCommandHandler(IUnitOfWork unitOfWork) : IRequestHandler<DeleteDoctorCommand, ErrorOr<Unit>>
+﻿public class DeleteDoctorCommandHandler(IUnitOfWork unitOfWork, IAccountHttpClient accountHttpClient) 
+    : IRequestHandler<DeleteDoctorCommand, ErrorOr<Unit>>
 {
     public async Task<ErrorOr<Unit>> Handle(DeleteDoctorCommand request, CancellationToken cancellationToken)
     {
@@ -10,6 +11,12 @@
             if (doctor == null)
             {
                 return Errors.Doctors.NotFound;
+            }
+
+            var doctorsAccountDeletionResponse = await accountHttpClient.DeleteAccount(doctor.AccountId);
+            if (doctorsAccountDeletionResponse.IsError)
+            {
+                return doctorsAccountDeletionResponse.FirstError;
             }
 
             await unitOfWork.DoctorsRepository.DeleteDoctorAsync(doctor.Id);
